% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IWLS.R
\name{cond_proposaldensity}
\alias{cond_proposaldensity}
\title{conditional proposal density q}
\usage{
cond_proposaldensity(x, y)
}
\arguments{
\item{x}{beta_t or}

\item{y}{beta_star}
}
\value{
q(beta_t | beta_star) or q(beta_star| beta_t)
}
\description{
conditional proposal density q
}
\examples{
cond_proposaldensity(beta_t, beta_star)
}
