% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis.R
\name{metrohas}
\alias{metrohas}
\title{Sample betas from a proposal density using MCMC}
\usage{
metrohas(startvalue, anzahl_sim)
}
\arguments{
\item{startvalue}{a beta vector}

\item{anzahl_sim}{number n of betas to sample, determines the length of the chain}
}
\value{
chain of n betas that will be ploted
 as a histogram or line plot to show the sampling path
}
\description{
The function executes the Metropolis Hashtings Algorithm, where a Normal proposal density
based on prior information and the current beta is generated (and updated) by means of iteratively
weighted least squares. A random sample from the proposal density is then evaluated according to
the loglikelihood, log-prior distribution and conditional proposal density compared to the previous
beta. If the new value performs good enough, it has a higher chance to be stored in the chain of
betas. If rejected, the current beta is stored in the chain and the algorithm proceeds to the next iteration.
}
\examples{
metrohas(c(1, 2, 3), 5000)
}
