% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IWLS.R
\name{proposalfunction}
\alias{proposalfunction}
\title{sample from IWLS proposal density}
\usage{
proposalfunction(beta_t)
}
\arguments{
\item{beta_t}{}
}
\value{
beta_star
}
\description{
sample from IWLS proposal density
}
\examples{
proposalfuntion(c(1, 2, 3))
}
