% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis.R
\name{metrohas}
\alias{metrohas}
\title{Sample betas from a proposal density using MCMC}
\usage{
metrohas(formula, dist, sigma2_start = 1, beta_start, a0 = 0.001,
  b0 = 1e-04, anzahl_sim, m = rep(0, length(beta_start)),
  M = diag(length(beta_start)), thinning_lag = 0)
}
\arguments{
\item{formula}{object of the type formula or a one that can be coerced into the class formula}

\item{anzahl_sim}{number n of betas to sample, determines the length of the chain}

\item{startvalue}{a beta vector}
}
\value{
chain of n betas that will be ploted
 as a histogram or line plot to show the sampling path
}
\description{
The function executes the Metropolis Hashtings Algorithm, where a Normal proposal density
based on prior information and the current beta is generated (and updated) by means of iteratively
weighted least squares. A random sample from the proposal density is then evaluated according to
the loglikelihood, log-prior distribution and conditional proposal density compared to the previous
beta. If the new value performs good enough, it has a higher chance to be stored in the chain of
betas. If rejected, the current beta is stored in the chain and the algorithm proceeds to the next iteration.
}
\examples{
metrohas(c(1, 2, 3), 5000)
}
