% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrohas.R
\name{summary.frequentistkiller}
\alias{summary.frequentistkiller}
\title{Title}
\usage{
\method{summary}{frequentistkiller}(result)
}
\arguments{
\item{result}{Object of class frequentistkiller}
}
\value{
list with all the elements of the summary output to the console
}
\description{
Title
}
