% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrohas.R
\name{summary.frequentistkiller}
\alias{summary.frequentistkiller}
\title{Title}
\usage{
\method{summary}{frequentistkiller}(object, ...)
}
\arguments{
\item{object}{list of S3 class "frequentistkiller"}

\item{...}{other arguments passed to the summary function, not aplicable for
this method}
}
\description{
Title
}
